push entrega final

+++ src/service/ProductService.js
@@ -1,29 +1,18 @@
 const ProductDAO = require('../dao/ProductDAO');
 const CustomError = require('./errors/CustomError');
 const ErrorCodes = require('./errors/errorCodes');
+const UserDAO = require('../dao/UserDAO')
+const transport = require('../utils/transport')
 const { generateInvalidProductDataError } = require('./errors')
 
 class ProductService {
-    async getProducts({ limit = 10, page = 1, sort, query } = {}) {
-        let filter = {};
-        let options = {
-            limit: parseInt(limit),
-            page: parseInt(page),
-            sort: {}
-        };
-
-        if (query) {
-            filter = { category: query };
+    async getAllProducts() {
+        try {
+            const products = await ProductDAO.find();
+            return products;
+        } catch (error) {
+            throw new Error('Error al extraer los productos');
         }
-
-        if (sort === 'asc' || sort === 'desc') {
-            options.sort.price = sort === 'asc' ? 1 : -1;
-        }
-
-        const result = await ProductDAO.paginate(filter, options);
-        const products = result.docs.map(doc => doc.toObject());
-
-        return { ...result, docs: products };
     }
 
     async getProductById(id) {
@@ -34,8 +23,8 @@ class ProductService {
         return product;
     }
 
-    async addProduct(title, description, price, thumbnail, code, status, stock, category) {
-        if (!title || !description || !code || !category || isNaN(+price) || +price <= 0 || isNaN(+stock) || +stock < 0) {
+    async addProduct(title, description, price, thumbnail, code, status, stock, category, owner) {
+        if (!title || !description || !code || !category || isNaN(+price) || +price <= 0 || isNaN(+stock) || +stock < 0 || !owner) {
             throw CustomError.customError({
                 name: 'datos invalidos',
                 message: 'Error al crear el producto, datos invalidos',
@@ -43,11 +32,11 @@ class ProductService {
                 code: ErrorCodes.INVALID_TYPES_ERROR
             });
         }
-
-        const finalThumbnail = thumbnail ? thumbnail : 'Sin Imagen';
+    
+        const finalThumbnail = thumbnail || 'Sin Imagen';
         const finalStatus = typeof status === 'undefined' || status === true || status === 'true';
-
-        await ProductDAO.Create({
+    
+        await ProductDAO.create({
             title,
             description,
             price,
@@ -55,9 +44,11 @@ class ProductService {
             code,
             status: finalStatus,
             stock,
-            category
+            category,
+            owner
         });
     }
+    
 
     async updateProduct(id, fieldsToUpdate) {
         if (Object.keys(fieldsToUpdate).length === 0) {
@@ -73,8 +64,47 @@ class ProductService {
     }
 
     async deleteProduct(id) {
-        await ProductDAO.deleteOne({ _id: id });
+            try {
+                const product = await ProductDAO.findById(id);
+        
+                if (!product) {
+                    throw new Error('El producto no existe');
+                }
+        
+                await ProductDAO.deleteOne({ _id: id });
+        
+                if (product.owner !== 'admin') {
+                    const user = await UserDAO.findUserByEmail(product.owner);
+        
+                    if (user && user.rol === 'premium') {
+                        try {
+                            await transport.sendMail({
+                                from: 'giorginiciro@gmail.com',
+                                to: user.email,
+                                subject: 'Notificación de eliminación de producto',
+                                html: `
+                                <div>
+                                    <p>Hola ${user.firstName},</p>
+                                    <p>El producto con ID ${id} ha sido eliminado de tu cuenta.</p>
+                                    <p>Si tienes alguna pregunta, no dudes en contactarnos.</p>
+                                    <p>Saludos,</p>
+                                    <p>El equipo de tu aplicación</p>
+                                </div>
+                                `,
+                            });
+                            console.log('Correo enviado exitosamente a:', user.email);
+                        } catch (error) {
+                            console.error('Error al enviar el correo:', error);
+                        }
+                    }
+                }
+        
+            } catch (error) {
+                console.error('Error al eliminar el producto:', error);
+                throw error;
+            }
+        }
+        
     }
-}
 
 module.exports = new ProductService();
